requirements/requirements.txt
django
requests

"""
Dev requirements
"""
black # code formatting
python-dotenv # environment variables
pytest
